# docs/generated/openapi_snapshot.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "QMEFTrainerStubFitBody": {
        "properties": {
          "hyperparams": {
            "additionalProperties": true,
            "description": "Forwarded into StubTorchMLIPTrainer.fit meta",
            "title": "Hyperparams",
            "type": "object"
          },
          "qmef": {
            "additionalProperties": true,
            "description": "Validated QMEFDataset JSON",
            "title": "Qmef",
            "type": "object"
          }
        },
        "required": [
          "qmef"
        ],
        "title": "QMEFTrainerStubFitBody",
        "type": "object"
      },
      "QMEFValidateBody": {
        "properties": {
          "qmef": {
            "additionalProperties": true,
            "description": "JSON object matching QMEFDataset (frames + optional provenance_yaml)",
            "title": "Qmef",
            "type": "object"
          }
        },
        "required": [
          "qmef"
        ],
        "title": "QMEFValidateBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "YamlPreviewBody": {
        "properties": {
          "experiment_yaml": {
            "description": "YAML to validate; returns computable list without running chemistry",
            "title": "Experiment Yaml",
            "type": "string"
          },
          "include_computables_rich": {
            "default": false,
            "description": "Optional parallel field: adds computables_rich (schema computables_rich_v1) without removing computable_abstract",
            "title": "Include Computables Rich",
            "type": "boolean"
          }
        },
        "required": [
          "experiment_yaml"
        ],
        "title": "YamlPreviewBody",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local API + SQLite queue for product workflows and reproducibility metadata.",
    "title": "qchem-stack",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/meta/computables-preview": {
      "post": {
        "operationId": "computables_preview_v1_meta_computables_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YamlPreviewBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Computables Preview V1 Meta Computables Preview Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Computables Preview",
        "tags": [
          "meta",
          "product"
        ]
      }
    },
...
    "/v1/meta/ml-md-trainer-stub-fit": {
      "post": {
        "operationId": "ml_md_trainer_stub_fit_v1_meta_ml_md_trainer_stub_fit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QMEFTrainerStubFitBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ml Md Trainer Stub Fit V1 Meta Ml Md Trainer Stub Fit Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ml Md Trainer Stub Fit",
        "tags": [
          "ml_md"
        ]
      }
    },
...
    "/v1/meta/qmef-validate": {
      "post": {
        "operationId": "qmef_validate_v1_meta_qmef_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QMEFValidateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Qmef Validate V1 Meta Qmef Validate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Qmef Validate",
        "tags": [
          "ml_md"
        ]
      }
    },
    "/v1/meta/queue-stats": {
      "get": {
        "operationId": "queue_stats_v1_meta_queue_stats_get",
        "parameters": [
          {
            "description": "SQLite path; default QCHEM_JOB_DB or temp",
            "in": "query",
            "name": "job_db_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "SQLite path; default QCHEM_JOB_DB or temp",
              "title": "Job Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Queue Stats V1 Meta Queue Stats Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Queue Stats",
        "tags": [
          "meta",
          "product"
        ]
      }
    },
    "/v1/meta/workflow-preview": {
      "post": {
        "operationId": "workflow_preview_v1_meta_workflow_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YamlPreviewBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Workflow Preview V1 Meta Workflow Preview Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workflow Preview",
        "tags": [
          "meta",
          "product",
          "product"
        ]
      }
    },
...
  },
...
}